requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
coloredlogs==15.0.1
ddddocr==1.4.7
exceptiongroup==1.0.4
Flask==2.2.2
flatbuffers==22.10.26
h11==0.14.0
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mpmath==1.2.1
numpy==1.23.4
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
outcome==1.2.0
packaging==21.3
Pillow==9.3.0
pip==22.3.1
pycparser==2.21
pyparsing==3.0.9
pyreadline3==3.4.1
PySocks==1.7.1
requests==2.28.1
selenium==4.6.0
setuptools==41.2.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.11.1
trio-websocket==0.9.2
trio==0.22.0
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.10.0
protobuf==3.20.1
mysql-connector-python==8.0.31